+++ bot/seasons/evergreen/game.py
@@ -13,7 +13,7 @@ from discord.ext.commands import Cog, Context, group
 from bot.bot import SeasonalBot
 from bot.constants import STAFF_ROLES, Tokens
 from bot.decorators import with_role
-from bot.pagination import ImagePaginator, LinePaginator
+from bot.utils.pagination import ImagePaginator, LinePaginator
 
 # Base URL of IGDB API
 BASE_URL = "https://api-v3.igdb.com"
