+++ bot/seasons/evergreen/movie.py
@@ -9,7 +9,7 @@ from discord import Embed
 from discord.ext.commands import Bot, Cog, Context, group
 
 from bot.constants import Tokens
-from bot.pagination import ImagePaginator
+from bot.utils.pagination import ImagePaginator
 
 # Define base URL of TMDB
 BASE_URL = "https://api.themoviedb.org/3/"
