+++ bot/seasons/evergreen/reddit.py
@@ -5,8 +5,7 @@ import discord
 from discord.ext import commands
 from discord.ext.commands.cooldowns import BucketType
 
-from bot.pagination import ImagePaginator
-
+from bot.utils.pagination import ImagePaginator
 
 log = logging.getLogger(__name__)
